+++ frontend/src/apis/index.ts
@@ -24,3 +24,27 @@ export const createConfigWithAuth = (tokens: Tokens | null) =>
 export const clientBasic = axios.create(defaultConfig);
 
 export const client = axios.create(createConfigWithAuth(tokens));
+
+client.interceptors.request.use(config => {
+  const tokens = zipgoLocalStorage.getTokens();
+
+  if (tokens) {
+    // eslint-disable-next-line no-param-reassign
+    config.headers.Authorization = `Bearer ${tokens.accessToken}`;
+  }
+
+  return config;
+});
+
+client.interceptors.response.use(
+  response => response,
+  async error => {
+    if (error.response.status === 401) {
+      alert('세션이 만료되었습니다.');
+
+      zipgoLocalStorage.clearAuth();
+    }
+
+    return Promise.reject(error);
+  },
+);
